requirements.txt
garminconnect>=0.2.22
stravalib>=1.5
anthropic>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
